Rename stack to record
